Finally managing to break the mapper file and quick documentation of the classes
